Move the tests away from the code
