+++ crates/std_detect/src/detect/mod.rs
@@ -135,10 +135,3 @@ pub fn features() -> impl Iterator<Item = (&'static str, bool)> {
     }
     impl_()
 }
-
-#[test]
-fn features_roundtrip() {
-    for (f, _) in features() {
-        let _ = Feature::from_str(f).unwrap();
-    }
-}
